feat(prep-004): Decompose BrainSettingsPage into modular panels
- Create ModelConfigPanel.tsx (643 lines) - routing, constraints, circuits, stats
- Create TokenAnalyticsPanel.tsx (587 lines) - usage analytics with charts
- Create RealtimeUsagePanel.tsx (270 lines) - live session tracking
- Create ApiKeyPanel.tsx (352 lines) - API key management
- Create RagConfigPanel.tsx (261 lines) - RAG configuration
- Reduce BrainSettingsPage.tsx from 2109 to 275 lines

Main page now orchestrates panels via props with shared state management.
All TypeScript checks pass.
