+++ frontend/src/features/routing/components/BrainSettingsPage.tsx
@@ -15,33 +15,8 @@
  * - Token usage and cost analytics
  */
 
+import { useState } from 'react';
 import { useQuery } from '@tanstack/react-query';
-import {
-  AreaChart,
-  Area,
-  BarChart,
-  Bar,
-  CartesianGrid,
-  XAxis,
-  YAxis,
-  ResponsiveContainer,
-  Tooltip,
-} from 'recharts';
-import {
-  CheckCircle2,
-  Database,
-  Download,
-  DollarSign,
-  Eye,
-  EyeOff,
-  Key,
-  Loader2,
-  RefreshCw,
-  Settings2,
-  TrendingUp,
-  XCircle,
-} from 'lucide-react';
-import { useEffect, useRef, useState } from 'react';
 
 import { Button } from '@/components/ui/button';
 import {
@@ -51,292 +26,16 @@ import {
   CardHeader,
   CardTitle,
 } from '@/components/ui/card';
-import { Input } from '@/components/ui/input';
-import { Label } from '@/components/ui/label';
-import { Switch } from '@/components/ui/switch';
-import {
-  Table,
-  TableBody,
-  TableCell,
-  TableHead,
-  TableHeader,
-  TableRow,
-} from '@/components/ui/table';
 import { Tabs, TabsContent, TabsList, TabsTrigger } from '@/components/ui/tabs';
-import {
-  Select,
-  SelectContent,
-  SelectItem,
-  SelectTrigger,
-  SelectValue,
-} from '@/components/ui/select';
 import { brainSettingsApi } from '@/features/routing/api/brainSettingsApi';
-import {
-  type DailyStatsResponse,
-  type ModelPricingResponse,
-  type ModelUsageResponse,
-  type RealtimeUsageEvent,
-} from '@/features/routing/api/brainSettingsApi';
-
-// BRAIN-035B-04: Real-time Usage Counter Hook
-interface ActiveSession {
-  session_id: string;
-  provider: string;
-  model_name: string;
-  input_tokens: number;
-  output_tokens: number;
-  total_tokens: number;
-  cost: number;
-  is_complete: boolean;
-}
-
-interface UseRealtimeUsageResult {
-  activeSessions: ActiveSession[];
-  currentSession: ActiveSession | null;
-  totalTokens: number;
-  totalCost: number;
-  isConnected: boolean;
-}
-
-function useRealtimeUsage(): UseRealtimeUsageResult {
-  const [activeSessions, setActiveSessions] = useState<ActiveSession[]>([]);
-  const [isConnected, setIsConnected] = useState(false);
-  const eventSourceRef = useRef<EventSource | null>(null);
-
-  useEffect(() => {
-    // Subscribe to real-time usage events
-    eventSourceRef.current = brainSettingsApi.streamRealtimeUsage(
-      (event: RealtimeUsageEvent) => {
-        if (event.type === 'session_start') {
-          setActiveSessions((prev) => [
-            ...prev,
-            {
-              session_id: event.session_id,
-              provider: event.provider,
-              model_name: event.model_name,
-              input_tokens: 0,
-              output_tokens: 0,
-              total_tokens: 0,
-              cost: 0,
-              is_complete: false,
-            },
-          ]);
-        } else if (event.type === 'token_update') {
-          setActiveSessions((prev) =>
-            prev.map((s) =>
-              s.session_id === event.session_id
-                ? {
-                    ...s,
-                    input_tokens: event.input_tokens,
-                    output_tokens: event.output_tokens,
-                    total_tokens: event.total_tokens,
-                    cost: event.cost,
-                  }
-                : s
-            )
-          );
-        } else if (event.type === 'session_complete') {
-          setActiveSessions((prev) =>
-            prev.map((s) =>
-              s.session_id === event.session_id ? { ...s, is_complete: true } : s
-            )
-          );
-          // Remove completed sessions after a delay
-          setTimeout(() => {
-            setActiveSessions((prev) =>
-              prev.filter((s) => s.session_id !== event.session_id)
-            );
-          }, 5000);
-        } else if (event.type === 'session_state') {
-          setActiveSessions((prev) => {
-            const existing = prev.find((s) => s.session_id === event.session_id);
-            if (existing) {
-              return prev.map((s) =>
-                s.session_id === event.session_id
-                  ? {
-                      ...s,
-                      input_tokens: event.input_tokens,
-                      output_tokens: event.output_tokens,
-                      total_tokens: event.total_tokens,
-                      cost: event.cost,
-                      is_complete: event.is_complete,
-                    }
-                  : s
-              );
-            }
-            return [
-              ...prev,
-              {
-                session_id: event.session_id,
-                provider: event.provider,
-                model_name: event.model_name,
-                input_tokens: event.input_tokens,
-                output_tokens: event.output_tokens,
-                total_tokens: event.total_tokens,
-                cost: event.cost,
-                is_complete: event.is_complete,
-              },
-            ];
-          });
-        }
-        setIsConnected(true);
-      },
-      (error) => {
-        console.error('Real-time usage error:', error);
-        setIsConnected(false);
-      }
-    );
-
-    return () => {
-      if (eventSourceRef.current) {
-        eventSourceRef.current.close();
-      }
-    };
-  }, []);
-
-  // Calculate totals across all active sessions
-  const totalTokens = activeSessions.reduce((sum, s) => sum + s.total_tokens, 0);
-  const totalCost = activeSessions.reduce((sum, s) => sum + s.cost, 0);
-
-  // Get the most recently updated active session
-  const currentSession =
-    activeSessions.length > 0
-      ? activeSessions.reduce((latest, s) =>
-          !latest || s.total_tokens > latest.total_tokens ? s : latest
-        )
-      : null;
-
-  return {
-    activeSessions,
-    currentSession,
-    totalTokens,
-    totalCost,
-    isConnected,
-  };
-}
-
-// BRAIN-035B-04: Real-time Usage Counter Component
-function RealtimeUsageCounter() {
-  const { activeSessions, currentSession, isConnected } = useRealtimeUsage();
-
-  // Don't show if no active sessions and not connected
-  if (!isConnected && activeSessions.length === 0) {
-    return null;
-  }
-
-  return (
-    <Card
-      className={isConnected && activeSessions.length > 0 ? 'border-primary/50' : ''}
-    >
-      <CardHeader className="pb-3">
-        <div className="flex items-center justify-between">
-          <div>
-            <CardTitle className="flex items-center gap-2 text-lg">
-              {activeSessions.length > 0 ? (
-                <>
-                  <Loader2 className="h-4 w-4 animate-spin text-primary" />
-                  Live Usage
-                </>
-              ) : (
-                'Real-time Counter'
-              )}
-            </CardTitle>
-            <CardDescription>
-              {activeSessions.length > 0
-                ? `Active generation${activeSessions.length > 1 ? `s (${activeSessions.length})` : ''}`
-                : isConnected
-                  ? 'Connected - waiting for generation...'
-                  : 'Connecting...'}
-            </CardDescription>
-          </div>
-          <div
-            className={`flex items-center gap-1.5 text-xs ${isConnected ? 'text-green-500' : 'text-muted-foreground'}`}
-          >
-            <div
-              className={`h-2 w-2 rounded-full ${isConnected ? 'animate-pulse bg-green-500' : 'bg-muted-foreground'}`}
-            />
-            {isConnected ? 'Live' : 'Offline'}
-          </div>
-        </div>
-      </CardHeader>
-      {activeSessions.length > 0 && (
-        <CardContent className="space-y-4">
-          {/* Current session summary */}
-          {currentSession && (
-            <div className="grid grid-cols-2 gap-4 md:grid-cols-4">
-              <div>
-                <p className="text-xs text-muted-foreground">Tokens</p>
-                <p className="text-xl font-semibold">
-                  {currentSession.total_tokens.toLocaleString()}
-                </p>
-                <p className="text-xs text-muted-foreground">
-                  {currentSession.input_tokens} in / {currentSession.output_tokens} out
-                </p>
-              </div>
-              <div>
-                <p className="text-xs text-muted-foreground">Cost</p>
-                <p className="text-xl font-semibold text-green-600">
-                  ${currentSession.cost.toFixed(4)}
-                </p>
-              </div>
-              <div>
-                <p className="text-xs text-muted-foreground">Model</p>
-                <p className="text-sm font-medium capitalize">
-                  {currentSession.provider}
-                </p>
-                <p className="text-xs text-muted-foreground">
-                  {currentSession.model_name}
-                </p>
-              </div>
-              <div>
-                <p className="text-xs text-muted-foreground">Status</p>
-                <p
-                  className={`text-sm font-medium ${currentSession.is_complete ? 'text-green-500' : 'text-primary'}`}
-                >
-                  {currentSession.is_complete ? 'Complete' : 'Generating...'}
-                </p>
-              </div>
-            </div>
-          )}
-
-          {/* All active sessions list */}
-          {activeSessions.length > 1 && (
-            <div className="space-y-2 border-t pt-2">
-              <p className="text-xs text-muted-foreground">All Active Sessions</p>
-              {activeSessions.map((session) => (
-                <div
-                  key={session.session_id}
-                  className="flex items-center justify-between rounded bg-muted/50 p-2"
-                >
-                  <div className="flex items-center gap-3">
-                    <div
-                      className={`h-2 w-2 rounded-full ${session.is_complete ? 'bg-green-500' : 'animate-pulse bg-primary'}`}
-                    />
-                    <div>
-                      <p className="text-sm font-medium capitalize">
-                        {session.provider}
-                      </p>
-                      <p className="text-xs text-muted-foreground">
-                        {session.model_name}
-                      </p>
-                    </div>
-                  </div>
-                  <div className="text-right">
-                    <p className="text-sm font-medium">
-                      {session.total_tokens.toLocaleString()} tokens
-                    </p>
-                    <p className="text-xs text-green-600">${session.cost.toFixed(4)}</p>
-                  </div>
-                </div>
-              ))}
-            </div>
-          )}
-        </CardContent>
-      )}
-    </Card>
-  );
-}
 import { routingApi } from '@/features/routing/api/routingApi';
+import {
+  ApiKeyPanel,
+  ModelConfigPanel,
+  RagConfigPanel,
+  TokenAnalyticsPanel,
+} from './panels';
+import { RefreshCw, Settings2 } from 'lucide-react';
 
 // Types for brain settings
 interface RAGConfigData {
@@ -423,20 +122,10 @@ const PROVIDER_MODELS: Record<Provider, string[]> = {
   mock: ['mock-model'],
 };
 
-export function BrainSettingsPage() {
-  // Simple toast notification helper (replace with useToast if available)
-  const toast = ({
-    title,
-    description,
-    variant = 'default',
-  }: {
-    title: string;
-    description?: string;
-    variant?: 'default';
-  }) => {
-    console.log(`[${variant.toUpperCase()}] ${title}: ${description}`);
-  };
+export { TASK_LABELS, TASK_DESCRIPTIONS, PROVIDER_LABELS, PROVIDER_MODELS };
+export type { TaskRule, RoutingConstraints, RoutingConfig, RAGConfigData };
 
+export function BrainSettingsPage() {
   const [isSaving, setIsSaving] = useState(false);
 
   // Fetch routing configuration
@@ -458,7 +147,7 @@ export function BrainSettingsPage() {
   } = useQuery({
     queryKey: ['routing-stats'],
     queryFn: () => routingApi.getStats(),
-    refetchInterval: 30000, // Refresh every 30 seconds
+    refetchInterval: 30000,
   });
 
   // BRAIN-033: Fetch brain settings data
@@ -471,306 +160,10 @@ export function BrainSettingsPage() {
     queryFn: () => brainSettingsApi.getSettings(),
   });
 
-  const [visibleKeys, setVisibleKeys] = useState<Record<string, boolean>>({});
-  const [localKeys, setLocalKeys] = useState<Record<string, string>>({
-    openai: '',
-    anthropic: '',
-    gemini: '',
-  });
-  const [localOllamaUrl, setLocalOllamaUrl] = useState('http://localhost:11434');
-
-  // BRAIN-035B-02: Date range filter state
-  const [dateRangeMode, setDateRangeMode] = useState<'preset' | 'custom'>('preset');
-  const [usageDays, setUsageDays] = useState(30);
-  const [customStartDate, setCustomStartDate] = useState<string>('');
-  const [customEndDate, setCustomEndDate] = useState<string>('');
-
-  // Calculate effective date range for queries
-  const getDaysFromDateRange = (): number => {
-    if (dateRangeMode === 'custom' && customStartDate && customEndDate) {
-      const start = new Date(customStartDate);
-      const end = new Date(customEndDate);
-      const diffTime = Math.abs(end.getTime() - start.getTime());
-      return Math.ceil(diffTime / (1000 * 60 * 60 * 24)) + 1;
-    }
-    return usageDays;
-  };
-
-  const effectiveDays = getDaysFromDateRange();
-
-  // BRAIN-035A: Fetch usage analytics data
-  const { data: usageSummary, isLoading: usageSummaryLoading } = useQuery({
-    queryKey: ['usage-summary', effectiveDays],
-    queryFn: () => brainSettingsApi.getUsageSummary(effectiveDays),
-  });
-
-  const { data: dailyUsage, isLoading: dailyUsageLoading } = useQuery({
-    queryKey: ['daily-usage', effectiveDays],
-    queryFn: () => brainSettingsApi.getDailyUsage(effectiveDays),
-  });
-
-  const { data: usageByModel, isLoading: usageByModelLoading } = useQuery({
-    queryKey: ['usage-by-model'],
-    queryFn: () => brainSettingsApi.getUsageByModel(),
-  });
-
-  // BRAIN-035B-01: Fetch model pricing data
-  const { data: modelPricing, isLoading: modelPricingLoading } = useQuery({
-    queryKey: ['model-pricing'],
-    queryFn: () => brainSettingsApi.getModelPricing(),
-  });
-
-  const handleSaveAPIKey = async (provider: string, key: string) => {
-    setIsSaving(true);
-    try {
-      await brainSettingsApi.updateAPIKeys({ [`${provider}_key`]: key || null });
-      await refetchBrainSettings();
-      setLocalKeys({ ...localKeys, [provider]: '' });
-      toast({
-        title: 'API Key saved',
-        description: `${provider.charAt(0).toUpperCase() + provider.slice(1)} API key has been updated.`,
-      });
-    } catch (error) {
-      toast({
-        title: 'Failed to save',
-        description: error instanceof Error ? error.message : 'Unknown error',
-      });
-    } finally {
-      setIsSaving(false);
-    }
-  };
-
-  const handleSaveOllamaUrl = async (url: string) => {
-    setIsSaving(true);
-    try {
-      await brainSettingsApi.updateAPIKeys({ ollama_base_url: url });
-      await refetchBrainSettings();
-      toast({
-        title: 'Ollama URL saved',
-        description: 'Ollama base URL has been updated.',
-      });
-    } catch (error) {
-      toast({
-        title: 'Failed to save',
-        description: error instanceof Error ? error.message : 'Unknown error',
-      });
-    } finally {
-      setIsSaving(false);
-    }
-  };
-
-  const handleSaveRAGConfig = async (updates: Partial<RAGConfigData>) => {
-    setIsSaving(true);
-    try {
-      await brainSettingsApi.updateRAGConfig(updates);
-      await refetchBrainSettings();
-      toast({
-        title: 'RAG configuration saved',
-        description: 'Your RAG settings have been updated.',
-      });
-    } catch (error) {
-      toast({
-        title: 'Failed to save',
-        description: error instanceof Error ? error.message : 'Unknown error',
-      });
-    } finally {
-      setIsSaving(false);
-    }
-  };
-
-  const handleSaveConfig = async (updates: Partial<RoutingConfig>) => {
-    if (!config) return;
-
-    setIsSaving(true);
-    try {
-      await routingApi.updateConfig(undefined, updates);
-      await refetchConfig();
-      toast({
-        title: 'Configuration saved',
-        description: 'Your routing preferences have been updated.',
-      });
-    } catch (error) {
-      toast({
-        title: 'Failed to save',
-        description: error instanceof Error ? error.message : 'Unknown error',
-      });
-    } finally {
-      setIsSaving(false);
-    }
-  };
-
-  const handleResetCircuitBreaker = async (modelKey: string) => {
-    try {
-      await routingApi.resetCircuitBreaker(modelKey);
-      await refetchStats();
-      toast({
-        title: 'Circuit breaker reset',
-        description: `Circuit breaker for ${modelKey} has been reset.`,
-      });
-    } catch (error) {
-      toast({
-        title: 'Failed to reset',
-        description: error instanceof Error ? error.message : 'Unknown error',
-      });
-    }
-  };
-
-  const TaskRuleCard = ({
-    rule,
-    taskType,
-  }: {
-    rule: TaskRule | undefined;
-    taskType: TaskType;
-  }) => {
-    const [localProvider, setLocalProvider] = useState<Provider>(
-      (rule?.provider as Provider) || 'gemini'
-    );
-    const [localModel, setLocalModel] = useState(rule?.model_name || '');
-    const [localTemp, setLocalTemp] = useState(rule?.temperature ?? null);
-    const [localMaxTokens, setLocalMaxTokens] = useState(rule?.max_tokens ?? null);
-    const [localEnabled, setLocalEnabled] = useState(rule?.enabled ?? true);
-
-    const availableModels = PROVIDER_MODELS[localProvider];
-
-    const handleSave = () => {
-      const updatedRules =
-        config?.task_rules.map((r) =>
-          r.task_type === taskType
-            ? {
-                ...r,
-                provider: localProvider,
-                model_name: localModel,
-                temperature: localTemp,
-                max_tokens: localMaxTokens,
-                enabled: localEnabled,
-              }
-            : r
-        ) || [];
-
-      const newRule: TaskRule = {
-        task_type: taskType,
-        provider: localProvider,
-        model_name: localModel,
-        temperature: localTemp,
-        max_tokens: localMaxTokens,
-        priority: rule?.priority || 0,
-        enabled: localEnabled,
-      };
-
-      const existingRuleIndex = updatedRules.findIndex((r) => r.task_type === taskType);
-      if (existingRuleIndex >= 0) {
-        updatedRules[existingRuleIndex] = newRule;
-      } else {
-        updatedRules.push(newRule);
-      }
-
-      handleSaveConfig({ task_rules: updatedRules });
-    };
-
-    return (
-      <Card className={!localEnabled ? 'opacity-60' : ''}>
-        <CardHeader className="pb-3">
-          <div className="flex items-center justify-between">
-            <div>
-              <CardTitle className="text-lg">{TASK_LABELS[taskType]}</CardTitle>
-              <CardDescription className="text-xs">
-                {TASK_DESCRIPTIONS[taskType]}
-              </CardDescription>
-            </div>
-            <Switch checked={localEnabled} onCheckedChange={setLocalEnabled} />
-          </div>
-        </CardHeader>
-        <CardContent className="space-y-4">
-          <div className="grid grid-cols-2 gap-4">
-            <div className="space-y-2">
-              <Label htmlFor={`provider-${taskType}`}>Provider</Label>
-              <Select
-                value={localProvider}
-                onValueChange={(v) => {
-                  setLocalProvider(v as Provider);
-                  setLocalModel('');
-                }}
-                disabled={!localEnabled}
-              >
-                <SelectTrigger id={`provider-${taskType}`}>
-                  <SelectValue />
-                </SelectTrigger>
-                <SelectContent>
-                  {Object.entries(PROVIDER_LABELS).map(([value, label]) => (
-                    <SelectItem key={value} value={value}>
-                      {label}
-                    </SelectItem>
-                  ))}
-                </SelectContent>
-              </Select>
-            </div>
-
-            <div className="space-y-2">
-              <Label htmlFor={`model-${taskType}`}>Model</Label>
-              <Select
-                value={localModel || 'default'}
-                onValueChange={(v) => setLocalModel(v === 'default' ? '' : v)}
-                disabled={!localEnabled || !localProvider}
-              >
-                <SelectTrigger id={`model-${taskType}`}>
-                  <SelectValue placeholder="Provider default" />
-                </SelectTrigger>
-                <SelectContent>
-                  <SelectItem value="default">Provider default</SelectItem>
-                  {availableModels.map((model) => (
-                    <SelectItem key={model} value={model}>
-                      {model}
-                    </SelectItem>
-                  ))}
-                </SelectContent>
-              </Select>
-            </div>
-          </div>
-
-          <div className="grid grid-cols-2 gap-4">
-            <div className="space-y-2">
-              <Label htmlFor={`temp-${taskType}`}>Temperature</Label>
-              <Input
-                id={`temp-${taskType}`}
-                type="number"
-                step="0.1"
-                min="0"
-                max="2"
-                placeholder="Default"
-                value={localTemp ?? ''}
-                onChange={(e) =>
-                  setLocalTemp(e.target.value ? parseFloat(e.target.value) : null)
-                }
-                disabled={!localEnabled}
-              />
-            </div>
-
-            <div className="space-y-2">
-              <Label htmlFor={`tokens-${taskType}`}>Max Tokens</Label>
-              <Input
-                id={`tokens-${taskType}`}
-                type="number"
-                min="1"
-                placeholder="Default"
-                value={localMaxTokens ?? ''}
-                onChange={(e) =>
-                  setLocalMaxTokens(
-                    e.target.value ? parseInt(e.target.value, 10) : null
-                  )
-                }
-                disabled={!localEnabled}
-              />
-            </div>
-          </div>
-
-          <div className="flex justify-end">
-            <Button size="sm" onClick={handleSave} disabled={!localEnabled || isSaving}>
-              {isSaving ? 'Saving...' : 'Save'}
-            </Button>
-          </div>
-        </CardContent>
-      </Card>
-    );
+  const handleRefetchAll = () => {
+    refetchConfig();
+    refetchStats();
+    refetchBrainSettings();
   };
 
   if (configLoading) {
@@ -821,7 +214,7 @@ export function BrainSettingsPage() {
           </p>
         </div>
         <div className="flex gap-2">
-          <Button variant="outline" onClick={() => refetchConfig()} size="sm">
+          <Button variant="outline" onClick={handleRefetchAll} size="sm">
             <RefreshCw className="mr-2 h-4 w-4" />
             Refresh
           </Button>
@@ -829,1276 +222,49 @@ export function BrainSettingsPage() {
       </div>
 
       <Tabs defaultValue="routing" className="space-y-6">
-        <TabsList className="grid w-full grid-cols-7 lg:w-auto">
-          <TabsTrigger value="routing">Model Routing</TabsTrigger>
-          <TabsTrigger value="constraints">Constraints</TabsTrigger>
-          <TabsTrigger value="circuits">Circuit Breakers</TabsTrigger>
-          <TabsTrigger value="stats">Statistics</TabsTrigger>
+        <TabsList className="grid w-full grid-cols-5 lg:w-auto">
+          <TabsTrigger value="routing">Model Config</TabsTrigger>
           <TabsTrigger value="usage">Usage</TabsTrigger>
           <TabsTrigger value="api-keys">API Keys</TabsTrigger>
           <TabsTrigger value="rag-settings">RAG Settings</TabsTrigger>
         </TabsList>
 
-        {/* Model Routing Tab */}
+        {/* Model Config Tab - Combines routing, constraints, circuits, and stats */}
         <TabsContent value="routing" className="space-y-6">
-          <Card>
-            <CardHeader>
-              <CardTitle>Global Settings</CardTitle>
-              <CardDescription>Apply to all routing decisions</CardDescription>
-            </CardHeader>
-            <CardContent className="flex gap-6">
-              <div className="flex items-center justify-between">
-                <Label htmlFor="enable-fallback">Enable Fallback Chain</Label>
-                <Switch
-                  id="enable-fallback"
-                  checked={config?.enable_fallback ?? true}
-                  onCheckedChange={(checked) =>
-                    handleSaveConfig({ enable_fallback: checked })
-                  }
-                />
-              </div>
-              <div className="flex items-center justify-between">
-                <Label htmlFor="enable-circuit">Enable Circuit Breaker</Label>
-                <Switch
-                  id="enable-circuit"
-                  checked={config?.enable_circuit_breaker ?? true}
-                  onCheckedChange={(checked) =>
-                    handleSaveConfig({ enable_circuit_breaker: checked })
-                  }
-                />
-              </div>
-            </CardContent>
-          </Card>
-
-          <div className="space-y-4">
-            <h2 className="text-xl font-semibold">Task-Based Routing</h2>
-            <p className="text-sm text-muted-foreground">
-              Configure which model to use for each type of task. Higher priority rules
-              take precedence.
-            </p>
-            <div className="grid gap-4 md:grid-cols-2">
-              {(['creative', 'logical', 'fast', 'cheap'] as TaskType[]).map(
-                (taskType) => {
-                  const rule = config?.task_rules.find((r) => r.task_type === taskType);
-                  return (
-                    <TaskRuleCard key={taskType} rule={rule} taskType={taskType} />
-                  );
-                }
-              )}
-            </div>
-          </div>
+          <ModelConfigPanel
+            config={config}
+            stats={stats}
+            statsLoading={statsLoading}
+            isSaving={isSaving}
+            onSaveComplete={handleRefetchAll}
+            onSavingChange={setIsSaving}
+          />
         </TabsContent>
 
-        {/* Constraints Tab */}
-        <TabsContent value="constraints" className="space-y-6">
-          <Card>
-            <CardHeader>
-              <CardTitle>Routing Constraints</CardTitle>
-              <CardDescription>
-                Limit which models can be selected based on cost, latency, and other
-                factors
-              </CardDescription>
-            </CardHeader>
-            <CardContent className="space-y-4">
-              <div className="space-y-2">
-                <Label htmlFor="max-cost">Maximum Cost (per 1M tokens)</Label>
-                <Input
-                  id="max-cost"
-                  type="number"
-                  step="0.01"
-                  min="0"
-                  placeholder="No limit"
-                  value={config?.constraints?.max_cost_per_1m_tokens ?? ''}
-                  onChange={(e) =>
-                    handleSaveConfig({
-                      constraints: {
-                        ...config?.constraints,
-                        max_cost_per_1m_tokens: e.target.value
-                          ? parseFloat(e.target.value)
-                          : null,
-                      } as RoutingConstraints,
-                    })
-                  }
-                />
-              </div>
-
-              <div className="space-y-2">
-                <Label htmlFor="max-latency">Maximum Latency (ms)</Label>
-                <Input
-                  id="max-latency"
-                  type="number"
-                  min="0"
-                  placeholder="No limit"
-                  value={config?.constraints?.max_latency_ms ?? ''}
-                  onChange={(e) =>
-                    handleSaveConfig({
-                      constraints: {
-                        ...config?.constraints,
-                        max_latency_ms: e.target.value
-                          ? parseInt(e.target.value, 10)
-                          : null,
-                      } as RoutingConstraints,
-                    })
-                  }
-                />
-              </div>
-            </CardContent>
-          </Card>
-
-          <Card>
-            <CardHeader>
-              <CardTitle>Provider Preferences</CardTitle>
-              <CardDescription>
-                Order providers by preference for routing decisions
-              </CardDescription>
-            </CardHeader>
-            <CardContent>
-              <div className="space-y-2">
-                {Object.entries(PROVIDER_LABELS).map(([value, label]) => {
-                  const isBlocked =
-                    config?.constraints?.blocked_providers?.includes(value);
-                  return (
-                    <div
-                      key={value}
-                      className="flex items-center justify-between rounded border p-2"
-                    >
-                      <span>{label}</span>
-                      <Switch
-                        checked={!isBlocked}
-                        onCheckedChange={(checked) => {
-                          const currentBlocked =
-                            config?.constraints?.blocked_providers || [];
-                          const newBlocked = checked
-                            ? currentBlocked.filter((p) => p !== value)
-                            : [...currentBlocked, value];
-                          handleSaveConfig({
-                            constraints: {
-                              ...config?.constraints,
-                              blocked_providers: newBlocked,
-                            } as RoutingConstraints,
-                          });
-                        }}
-                      />
-                    </div>
-                  );
-                })}
-              </div>
-            </CardContent>
-          </Card>
-        </TabsContent>
-
-        {/* Circuit Breakers Tab */}
-        <TabsContent value="circuits" className="space-y-6">
-          <Card>
-            <CardHeader>
-              <CardTitle>Circuit Breaker Status</CardTitle>
-              <CardDescription>
-                Circuit breakers prevent cascading failures by temporarily disabling
-                failing models
-              </CardDescription>
-            </CardHeader>
-            <CardContent>
-              {statsLoading ? (
-                <div className="h-32 animate-pulse rounded bg-muted" />
-              ) : stats?.open_circuits && stats.open_circuits.length > 0 ? (
-                <div className="space-y-2">
-                  {stats.open_circuits.map((circuit) => (
-                    <div
-                      key={circuit.model}
-                      className="destructive/10 flex items-center justify-between rounded border border-destructive p-3"
-                    >
-                      <div>
-                        <div className="font-medium">{circuit.model}</div>
-                        <div className="text-sm text-muted-foreground">
-                          Failures: {circuit.failure_count} | State: {circuit.state}
-                        </div>
-                      </div>
-                      <Button
-                        size="sm"
-                        variant="outline"
-                        onClick={() => handleResetCircuitBreaker(circuit.model)}
-                      >
-                        Reset
-                      </Button>
-                    </div>
-                  ))}
-                </div>
-              ) : (
-                <p className="text-sm text-muted-foreground">
-                  No open circuits. All models are healthy.
-                </p>
-              )}
-            </CardContent>
-          </Card>
-
-          <Card>
-            <CardHeader>
-              <CardTitle>Circuit Breaker Configuration</CardTitle>
-              <CardDescription>
-                Configure thresholds and timeouts for specific models
-              </CardDescription>
-            </CardHeader>
-            <CardContent>
-              <p className="text-sm text-muted-foreground">
-                Circuit breaker rules can be configured per model. Default: 5 failures
-                triggers timeout, 60 seconds before retry.
-              </p>
-            </CardContent>
-          </Card>
-        </TabsContent>
-
-        {/* Statistics Tab */}
-        <TabsContent value="stats" className="space-y-6">
-          <div className="grid gap-4 md:grid-cols-4">
-            <Card>
-              <CardHeader className="pb-2">
-                <CardDescription>Total Decisions</CardDescription>
-                <CardTitle className="text-2xl">
-                  {statsLoading
-                    ? '...'
-                    : (stats?.total_decisions.toLocaleString() ?? 0)}
-                </CardTitle>
-              </CardHeader>
-            </Card>
-            <Card>
-              <CardHeader className="pb-2">
-                <CardDescription>Fallback Rate</CardDescription>
-                <CardTitle className="text-2xl">
-                  {statsLoading
-                    ? '...'
-                    : `${((stats?.fallback_rate ?? 0) * 100).toFixed(1)}%`}
-                </CardTitle>
-              </CardHeader>
-            </Card>
-            <Card>
-              <CardHeader className="pb-2">
-                <CardDescription>Avg Routing Time</CardDescription>
-                <CardTitle className="text-2xl">
-                  {statsLoading ? '...' : `${stats?.avg_routing_time_ms.toFixed(2)}ms`}
-                </CardTitle>
-              </CardHeader>
-            </Card>
-            <Card>
-              <CardHeader className="pb-2">
-                <CardDescription>Active Circuits</CardDescription>
-                <CardTitle className="text-2xl">
-                  {statsLoading ? '...' : (stats?.open_circuits.length ?? 0)}
-                </CardTitle>
-              </CardHeader>
-            </Card>
-          </div>
-
-          <Card>
-            <CardHeader>
-              <CardTitle>Provider Usage</CardTitle>
-              <CardDescription>
-                Number of requests routed to each provider
-              </CardDescription>
-            </CardHeader>
-            <CardContent>
-              {statsLoading ? (
-                <div className="h-32 animate-pulse rounded bg-muted" />
-              ) : (
-                <div className="space-y-2">
-                  {Object.entries(stats?.provider_counts || {}).map(
-                    ([provider, count]) => (
-                      <div key={provider} className="flex items-center gap-2">
-                        <span className="w-24 text-sm">
-                          {PROVIDER_LABELS[provider as Provider] || provider}
-                        </span>
-                        <div className="h-2 flex-1 rounded-full bg-muted">
-                          <div
-                            className="h-2 rounded-full bg-primary"
-                            style={{
-                              width: `${(count / (stats?.total_decisions || 1)) * 100}%`,
-                            }}
-                          />
-                        </div>
-                        <span className="w-12 text-right text-sm">{count}</span>
-                      </div>
-                    )
-                  )}
-                </div>
-              )}
-            </CardContent>
-          </Card>
-
-          <Card>
-            <CardHeader>
-              <CardTitle>Routing Reasons</CardTitle>
-              <CardDescription>Why models were selected for routing</CardDescription>
-            </CardHeader>
-            <CardContent>
-              {statsLoading ? (
-                <div className="h-32 animate-pulse rounded bg-muted" />
-              ) : (
-                <div className="space-y-2">
-                  {Object.entries(stats?.reason_counts || {}).map(([reason, count]) => (
-                    <div
-                      key={reason}
-                      className="flex items-center justify-between rounded border p-2"
-                    >
-                      <span className="text-sm capitalize">
-                        {reason.replace(/_/g, ' ')}
-                      </span>
-                      <span className="font-medium">{count}</span>
-                    </div>
-                  ))}
-                </div>
-              )}
-            </CardContent>
-          </Card>
-        </TabsContent>
-
-        {/* BRAIN-035A: Usage Analytics Tab */}
+        {/* Usage Analytics Tab */}
         <TabsContent value="usage" className="space-y-6">
-          {/* BRAIN-035B-04: Real-time Usage Counter */}
-          <RealtimeUsageCounter />
-
-          {/* Summary Stats */}
-          <div className="grid gap-4 md:grid-cols-4">
-            <Card>
-              <CardHeader className="pb-2">
-                <CardDescription>Total Tokens</CardDescription>
-                <CardTitle className="flex items-center gap-2 text-2xl">
-                  {usageSummaryLoading ? (
-                    <Loader2 className="h-5 w-5 animate-spin" />
-                  ) : (
-                    <>
-                      <TrendingUp className="h-5 w-5 text-primary" />
-                      {(usageSummary?.total_tokens ?? 0).toLocaleString()}
-                    </>
-                  )}
-                </CardTitle>
-              </CardHeader>
-              <CardContent className="pt-2">
-                <p className="text-xs text-muted-foreground">
-                  {usageSummary?.period_end ? `Last ${usageDays} days` : '-'}
-                </p>
-              </CardContent>
-            </Card>
-            <Card>
-              <CardHeader className="pb-2">
-                <CardDescription>Total Cost</CardDescription>
-                <CardTitle className="flex items-center gap-2 text-2xl">
-                  {usageSummaryLoading ? (
-                    <Loader2 className="h-5 w-5 animate-spin" />
-                  ) : (
-                    <>
-                      <DollarSign className="h-5 w-5 text-green-500" />$
-                      {(usageSummary?.total_cost ?? 0).toFixed(4)}
-                    </>
-                  )}
-                </CardTitle>
-              </CardHeader>
-              <CardContent className="pt-2">
-                <p className="text-xs text-muted-foreground">
-                  {usageSummary?.period_end ? `Last ${usageDays} days` : '-'}
-                </p>
-              </CardContent>
-            </Card>
-            <Card>
-              <CardHeader className="pb-2">
-                <CardDescription>Total Requests</CardDescription>
-                <CardTitle className="text-2xl">
-                  {usageSummaryLoading ? (
-                    <Loader2 className="h-5 w-5 animate-spin" />
-                  ) : (
-                    (usageSummary?.total_requests ?? 0).toLocaleString()
-                  )}
-                </CardTitle>
-              </CardHeader>
-              <CardContent className="pt-2">
-                <p className="text-xs text-muted-foreground">
-                  {usageSummary?.period_end
-                    ? `Avg ${((usageSummary?.total_tokens ?? 0) / (usageSummary?.total_requests || 1) || 0).toFixed(0)} tokens/request`
-                    : '-'}
-                </p>
-              </CardContent>
-            </Card>
-            <Card>
-              <CardHeader className="pb-2">
-                <CardDescription>Avg Latency</CardDescription>
-                <CardTitle className="text-2xl">
-                  {usageSummaryLoading ? (
-                    <Loader2 className="h-5 w-5 animate-spin" />
-                  ) : (
-                    `${(usageSummary?.avg_latency_ms ?? 0).toFixed(0)}ms`
-                  )}
-                </CardTitle>
-              </CardHeader>
-              <CardContent className="pt-2">
-                <p className="text-xs text-muted-foreground">
-                  {usageSummary?.period_end ? 'Across all requests' : '-'}
-                </p>
-              </CardContent>
-            </Card>
-          </div>
-
-          {/* Tokens Over Time Chart */}
-          <Card>
-            <CardHeader>
-              <div className="flex items-center justify-between">
-                <div>
-                  <CardTitle>Tokens Over Time</CardTitle>
-                  <CardDescription>Daily token usage trend</CardDescription>
-                </div>
-                <div className="flex items-center gap-4">
-                  {/* BRAIN-035B-02: Date range filter */}
-                  <div className="flex items-center gap-2">
-                    <Button
-                      size="sm"
-                      variant={dateRangeMode === 'preset' ? 'default' : 'outline'}
-                      onClick={() => setDateRangeMode('preset')}
-                    >
-                      Preset
-                    </Button>
-                    <Button
-                      size="sm"
-                      variant={dateRangeMode === 'custom' ? 'default' : 'outline'}
-                      onClick={() => setDateRangeMode('custom')}
-                    >
-                      Custom
-                    </Button>
-                  </div>
-
-                  {dateRangeMode === 'preset' ? (
-                    <div className="flex gap-2">
-                      {[7, 30, 90].map((days) => (
-                        <Button
-                          key={days}
-                          size="sm"
-                          variant={usageDays === days ? 'default' : 'outline'}
-                          onClick={() => setUsageDays(days)}
-                        >
-                          {days}d
-                        </Button>
-                      ))}
-                    </div>
-                  ) : (
-                    <div className="flex items-center gap-2">
-                      <Input
-                        type="date"
-                        value={customStartDate}
-                        onChange={(e) => setCustomStartDate(e.target.value)}
-                        className="w-auto"
-                      />
-                      <span className="text-muted-foreground">to</span>
-                      <Input
-                        type="date"
-                        value={customEndDate}
-                        onChange={(e) => setCustomEndDate(e.target.value)}
-                        className="w-auto"
-                      />
-                      <Button
-                        size="sm"
-                        disabled={!customStartDate || !customEndDate}
-                        onClick={() => {
-                          if (customStartDate && customEndDate) {
-                            const start = new Date(customStartDate);
-                            const end = new Date(customEndDate);
-                            const diffTime = Math.abs(end.getTime() - start.getTime());
-                            const days =
-                              Math.ceil(diffTime / (1000 * 60 * 60 * 24)) + 1;
-                            setUsageDays(days);
-                          }
-                        }}
-                      >
-                        Apply
-                      </Button>
-                    </div>
-                  )}
-
-                  {/* BRAIN-035B-03: CSV Export button */}
-                  <Button
-                    size="sm"
-                    variant="outline"
-                    onClick={() => brainSettingsApi.exportUsageCsv(effectiveDays)}
-                    className="gap-2"
-                  >
-                    <Download className="h-4 w-4" />
-                    Export CSV
-                  </Button>
-                </div>
-              </div>
-            </CardHeader>
-            <CardContent>
-              {dailyUsageLoading ? (
-                <div className="flex h-64 items-center justify-center">
-                  <Loader2 className="h-8 w-8 animate-spin text-muted-foreground" />
-                </div>
-              ) : dailyUsage && dailyUsage.length > 0 ? (
-                <div className="h-64">
-                  <ResponsiveContainer width="100%" height="100%">
-                    <AreaChart data={dailyUsage}>
-                      <defs>
-                        <linearGradient id="tokensGradient" x1="0" y1="0" x2="0" y2="1">
-                          <stop
-                            offset="5%"
-                            stopColor="hsl(var(--primary))"
-                            stopOpacity={0.3}
-                          />
-                          <stop
-                            offset="95%"
-                            stopColor="hsl(var(--primary))"
-                            stopOpacity={0}
-                          />
-                        </linearGradient>
-                      </defs>
-                      <CartesianGrid
-                        strokeDasharray="3 3"
-                        stroke="hsl(var(--border))"
-                        vertical={false}
-                      />
-                      <XAxis
-                        dataKey="date"
-                        tickFormatter={(value) => {
-                          const date = new Date(value);
-                          return date.toLocaleDateString('en-US', {
-                            month: 'short',
-                            day: 'numeric',
-                          });
-                        }}
-                        tick={{ fill: 'hsl(var(--muted-foreground))', fontSize: 12 }}
-                        tickLine={{ stroke: 'hsl(var(--border))' }}
-                        axisLine={{ stroke: 'hsl(var(--border))' }}
-                      />
-                      <YAxis
-                        tick={{ fill: 'hsl(var(--muted-foreground))', fontSize: 12 }}
-                        tickLine={{ stroke: 'hsl(var(--border))' }}
-                        axisLine={{ stroke: 'hsl(var(--border))' }}
-                      />
-                      <Tooltip
-                        content={({ active, payload }) => {
-                          if (!active || !payload?.[0]) return null;
-                          const data = payload[0].payload as DailyStatsResponse;
-                          return (
-                            <div className="rounded-md border bg-popover px-3 py-2 shadow-md">
-                              <p className="mb-1 font-medium">
-                                {new Date(data.date).toLocaleDateString('en-US', {
-                                  month: 'short',
-                                  day: 'numeric',
-                                })}
-                              </p>
-                              <p className="text-sm">
-                                Tokens:{' '}
-                                <span className="font-medium">
-                                  {data.total_tokens.toLocaleString()}
-                                </span>
-                              </p>
-                              <p className="text-sm">
-                                Cost:{' '}
-                                <span className="font-medium">
-                                  ${data.total_cost.toFixed(4)}
-                                </span>
-                              </p>
-                              <p className="text-sm">
-                                Requests:{' '}
-                                <span className="font-medium">
-                                  {data.total_requests}
-                                </span>
-                              </p>
-                            </div>
-                          );
-                        }}
-                      />
-                      <Area
-                        type="monotone"
-                        dataKey="total_tokens"
-                        stroke="hsl(var(--primary))"
-                        fill="url(#tokensGradient)"
-                        strokeWidth={2}
-                      />
-                    </AreaChart>
-                  </ResponsiveContainer>
-                </div>
-              ) : (
-                <div className="flex h-64 items-center justify-center text-muted-foreground">
-                  No usage data available
-                </div>
-              )}
-            </CardContent>
-          </Card>
-
-          {/* Cost by Model Chart */}
-          <Card>
-            <CardHeader>
-              <CardTitle>Cost by Model</CardTitle>
-              <CardDescription>Total cost breakdown per model</CardDescription>
-            </CardHeader>
-            <CardContent>
-              {usageByModelLoading ? (
-                <div className="flex h-64 items-center justify-center">
-                  <Loader2 className="h-8 w-8 animate-spin text-muted-foreground" />
-                </div>
-              ) : usageByModel && usageByModel.length > 0 ? (
-                <div className="h-64">
-                  <ResponsiveContainer width="100%" height="100%">
-                    <BarChart data={usageByModel}>
-                      <CartesianGrid
-                        strokeDasharray="3 3"
-                        stroke="hsl(var(--border))"
-                        vertical={false}
-                      />
-                      <XAxis
-                        dataKey="model_identifier"
-                        angle={-45}
-                        textAnchor="end"
-                        height={80}
-                        tick={{ fill: 'hsl(var(--muted-foreground))', fontSize: 11 }}
-                        tickLine={{ stroke: 'hsl(var(--border))' }}
-                        axisLine={{ stroke: 'hsl(var(--border))' }}
-                      />
-                      <YAxis
-                        tick={{ fill: 'hsl(var(--muted-foreground))', fontSize: 12 }}
-                        tickLine={{ stroke: 'hsl(var(--border))' }}
-                        axisLine={{ stroke: 'hsl(var(--border))' }}
-                      />
-                      <Tooltip
-                        content={({ active, payload }) => {
-                          if (!active || !payload?.[0]) return null;
-                          const data = payload[0].payload as ModelUsageResponse;
-                          return (
-                            <div className="rounded-md border bg-popover px-3 py-2 shadow-md">
-                              <p className="mb-1 font-medium">
-                                {data.model_identifier}
-                              </p>
-                              <p className="text-sm">
-                                Cost:{' '}
-                                <span className="font-medium">
-                                  ${data.total_cost.toFixed(4)}
-                                </span>
-                              </p>
-                              <p className="text-sm">
-                                Tokens:{' '}
-                                <span className="font-medium">
-                                  {data.total_tokens.toLocaleString()}
-                                </span>
-                              </p>
-                              <p className="text-sm">
-                                Requests:{' '}
-                                <span className="font-medium">
-                                  {data.total_requests}
-                                </span>
-                              </p>
-                            </div>
-                          );
-                        }}
-                      />
-                      <Bar
-                        dataKey="total_cost"
-                        fill="hsl(var(--primary))"
-                        radius={[4, 4, 0, 0]}
-                      />
-                    </BarChart>
-                  </ResponsiveContainer>
-                </div>
-              ) : (
-                <div className="flex h-64 items-center justify-center text-muted-foreground">
-                  No usage data available
-                </div>
-              )}
-            </CardContent>
-          </Card>
-
-          {/* Provider Distribution */}
-          <Card>
-            <CardHeader>
-              <CardTitle>Provider Distribution</CardTitle>
-              <CardDescription>Usage breakdown by LLM provider</CardDescription>
-            </CardHeader>
-            <CardContent>
-              {usageByModelLoading ? (
-                <div className="flex h-48 items-center justify-center">
-                  <Loader2 className="h-6 w-6 animate-spin text-muted-foreground" />
-                </div>
-              ) : usageByModel && usageByModel.length > 0 ? (
-                <div className="space-y-3">
-                  {Object.entries(
-                    usageByModel.reduce(
-                      (acc, item) => {
-                        const provider = item.provider;
-                        if (!acc[provider]) {
-                          acc[provider] = { cost: 0, tokens: 0, requests: 0 };
-                        }
-                        acc[provider].cost += item.total_cost;
-                        acc[provider].tokens += item.total_tokens;
-                        acc[provider].requests += item.total_requests;
-                        return acc;
-                      },
-                      {} as Record<
-                        string,
-                        { cost: number; tokens: number; requests: number }
-                      >
-                    )
-                  ).map(([provider, stats]) => {
-                    const totalCost = usageByModel.reduce(
-                      (sum, item) => sum + item.total_cost,
-                      0
-                    );
-                    const percentage =
-                      totalCost > 0 ? (stats.cost / totalCost) * 100 : 0;
-                    return (
-                      <div key={provider} className="space-y-1">
-                        <div className="flex items-center justify-between text-sm">
-                          <span className="font-medium capitalize">{provider}</span>
-                          <span className="text-muted-foreground">
-                            ${stats.cost.toFixed(4)} ({percentage.toFixed(1)}%)
-                          </span>
-                        </div>
-                        <div className="h-2 overflow-hidden rounded-full bg-muted">
-                          <div
-                            className="h-full rounded-full bg-primary"
-                            style={{ width: `${percentage}%` }}
-                          />
-                        </div>
-                      </div>
-                    );
-                  })}
-                </div>
-              ) : (
-                <div className="flex h-48 items-center justify-center text-muted-foreground">
-                  No usage data available
-                </div>
-              )}
-            </CardContent>
-          </Card>
-
-          {/* BRAIN-035B-01: Model Comparison Table */}
-          <Card>
-            <CardHeader>
-              <CardTitle>Model Pricing Comparison</CardTitle>
-              <CardDescription>
-                Cost per 1M tokens for each available model
-              </CardDescription>
-            </CardHeader>
-            <CardContent>
-              {modelPricingLoading ? (
-                <div className="flex h-48 items-center justify-center">
-                  <Loader2 className="h-6 w-6 animate-spin text-muted-foreground" />
-                </div>
-              ) : modelPricing && modelPricing.length > 0 ? (
-                <div className="space-y-6">
-                  {(() => {
-                    // Group models by provider
-                    const grouped = modelPricing.reduce(
-                      (acc, model) => {
-                        const provider = model.provider ?? 'unknown';
-                        if (!acc[provider]) {
-                          acc[provider] = [];
-                        }
-                        acc[provider].push(model);
-                        return acc;
-                      },
-                      {} as Record<string, ModelPricingResponse[]>
-                    );
-
-                    const providerLabels: Record<string, string> = {
-                      openai: 'OpenAI',
-                      anthropic: 'Anthropic',
-                      gemini: 'Google Gemini',
-                      ollama: 'Ollama (Local)',
-                      mock: 'Mock',
-                    };
-
-                    return Object.entries(grouped).map(([provider, models]) => (
-                      <div key={provider} className="space-y-3">
-                        <h3 className="flex items-center gap-2 text-lg font-semibold capitalize">
-                          {providerLabels[provider] || provider}
-                        </h3>
-                        <div className="overflow-x-auto rounded-lg border">
-                          <Table>
-                            <TableHeader>
-                              <TableRow>
-                                <TableHead>Model</TableHead>
-                                <TableHead className="text-right">Input / 1M</TableHead>
-                                <TableHead className="text-right">
-                                  Output / 1M
-                                </TableHead>
-                                <TableHead className="text-right">Context</TableHead>
-                              </TableRow>
-                            </TableHeader>
-                            <TableBody>
-                              {models.map((model) => (
-                                <TableRow key={`${model.provider}-${model.model_name}`}>
-                                  <TableCell>
-                                    <div>
-                                      <div className="font-medium">
-                                        {model.display_name}
-                                      </div>
-                                      <div className="text-xs text-muted-foreground">
-                                        {model.model_name}
-                                      </div>
-                                    </div>
-                                  </TableCell>
-                                  <TableCell className="text-right">
-                                    ${model.cost_per_1m_input_tokens.toFixed(2)}
-                                  </TableCell>
-                                  <TableCell className="text-right">
-                                    ${model.cost_per_1m_output_tokens.toFixed(2)}
-                                  </TableCell>
-                                  <TableCell className="text-right">
-                                    {model.max_context_tokens.toLocaleString()}
-                                  </TableCell>
-                                </TableRow>
-                              ))}
-                            </TableBody>
-                          </Table>
-                        </div>
-                      </div>
-                    ));
-                  })()}
-                </div>
-              ) : (
-                <div className="flex h-48 items-center justify-center text-muted-foreground">
-                  No model pricing data available
-                </div>
-              )}
-            </CardContent>
-          </Card>
+          <TokenAnalyticsPanel />
         </TabsContent>
 
-        {/* BRAIN-033: API Keys Tab */}
+        {/* API Keys Tab */}
         <TabsContent value="api-keys" className="space-y-6">
-          <Card>
-            <CardHeader>
-              <CardTitle className="flex items-center gap-2">
-                <Key className="h-5 w-5" />
-                API Keys
-              </CardTitle>
-              <CardDescription>
-                Configure API keys for LLM providers. Keys are encrypted and stored
-                securely.
-              </CardDescription>
-            </CardHeader>
-            <CardContent className="space-y-6">
-              {/* Knowledge Base Status */}
-              <div className="rounded-lg border bg-muted/50 p-4">
-                <h3 className="mb-3 flex items-center gap-2 font-semibold">
-                  <Database className="h-4 w-4" />
-                  Knowledge Base Status
-                </h3>
-                <div className="grid grid-cols-2 gap-4 md:grid-cols-5">
-                  <div>
-                    <div className="text-2xl font-bold">
-                      {brainSettingsLoading
-                        ? '...'
-                        : (brainSettings?.knowledge_base.total_entries ?? 0)}
-                    </div>
-                    <div className="text-xs text-muted-foreground">Total Entries</div>
-                  </div>
-                  <div>
-                    <div className="text-2xl font-bold">
-                      {brainSettingsLoading
-                        ? '...'
-                        : (brainSettings?.knowledge_base.characters_count ?? 0)}
-                    </div>
-                    <div className="text-xs text-muted-foreground">Characters</div>
-                  </div>
-                  <div>
-                    <div className="text-2xl font-bold">
-                      {brainSettingsLoading
-                        ? '...'
-                        : (brainSettings?.knowledge_base.lore_count ?? 0)}
-                    </div>
-                    <div className="text-xs text-muted-foreground">Lore</div>
-                  </div>
-                  <div>
-                    <div className="text-2xl font-bold">
-                      {brainSettingsLoading
-                        ? '...'
-                        : (brainSettings?.knowledge_base.scenes_count ?? 0)}
-                    </div>
-                    <div className="text-xs text-muted-foreground">Scenes</div>
-                  </div>
-                  <div>
-                    <div className="text-2xl font-bold">
-                      {brainSettings?.knowledge_base.is_healthy ? (
-                        <CheckCircle2 className="h-6 w-6 text-green-500" />
-                      ) : (
-                        <XCircle className="h-6 w-6 text-red-500" />
-                      )}
-                    </div>
-                    <div className="text-xs text-muted-foreground">Health</div>
-                  </div>
-                </div>
-              </div>
-
-              {/* OpenAI API Key */}
-              <div className="space-y-3">
-                <Label htmlFor="openai-key" className="flex items-center gap-2">
-                  <span>OpenAI API Key</span>
-                  {brainSettings?.api_keys.has_openai && (
-                    <CheckCircle2 className="h-4 w-4 text-green-500" />
-                  )}
-                </Label>
-                <div className="flex gap-2">
-                  <div className="relative flex-1">
-                    <Input
-                      id="openai-key"
-                      type={visibleKeys.openai ? 'text' : 'password'}
-                      placeholder={
-                        brainSettings?.api_keys.has_openai
-                          ? brainSettings.api_keys.openai_key
-                          : 'sk-...'
-                      }
-                      value={localKeys.openai || ''}
-                      onChange={(e) =>
-                        setLocalKeys({ ...localKeys, openai: e.target.value || '' })
-                      }
-                      className="pr-20"
-                    />
-                    <Button
-                      type="button"
-                      variant="ghost"
-                      size="sm"
-                      className="absolute right-1 top-1/2 h-7 -translate-y-1/2 px-2"
-                      onClick={() =>
-                        setVisibleKeys({ ...visibleKeys, openai: !visibleKeys.openai })
-                      }
-                    >
-                      {visibleKeys.openai ? (
-                        <EyeOff className="h-4 w-4" />
-                      ) : (
-                        <Eye className="h-4 w-4" />
-                      )}
-                    </Button>
-                  </div>
-                  <Button
-                    size="sm"
-                    onClick={() => handleSaveAPIKey('openai', localKeys.openai || '')}
-                    disabled={!localKeys.openai || isSaving}
-                  >
-                    Save
-                  </Button>
-                </div>
-              </div>
-
-              {/* Anthropic API Key */}
-              <div className="space-y-3">
-                <Label htmlFor="anthropic-key" className="flex items-center gap-2">
-                  <span>Anthropic API Key</span>
-                  {brainSettings?.api_keys.has_anthropic && (
-                    <CheckCircle2 className="h-4 w-4 text-green-500" />
-                  )}
-                </Label>
-                <div className="flex gap-2">
-                  <div className="relative flex-1">
-                    <Input
-                      id="anthropic-key"
-                      type={visibleKeys.anthropic ? 'text' : 'password'}
-                      placeholder={
-                        brainSettings?.api_keys.has_anthropic
-                          ? brainSettings.api_keys.anthropic_key
-                          : 'sk-ant-...'
-                      }
-                      value={localKeys.anthropic || ''}
-                      onChange={(e) =>
-                        setLocalKeys({ ...localKeys, anthropic: e.target.value || '' })
-                      }
-                      className="pr-20"
-                    />
-                    <Button
-                      type="button"
-                      variant="ghost"
-                      size="sm"
-                      className="absolute right-1 top-1/2 h-7 -translate-y-1/2 px-2"
-                      onClick={() =>
-                        setVisibleKeys({
-                          ...visibleKeys,
-                          anthropic: !visibleKeys.anthropic,
-                        })
-                      }
-                    >
-                      {visibleKeys.anthropic ? (
-                        <EyeOff className="h-4 w-4" />
-                      ) : (
-                        <Eye className="h-4 w-4" />
-                      )}
-                    </Button>
-                  </div>
-                  <Button
-                    size="sm"
-                    onClick={() =>
-                      handleSaveAPIKey('anthropic', localKeys.anthropic || '')
-                    }
-                    disabled={!localKeys.anthropic || isSaving}
-                  >
-                    Save
-                  </Button>
-                </div>
-              </div>
-
-              {/* Gemini API Key */}
-              <div className="space-y-3">
-                <Label htmlFor="gemini-key" className="flex items-center gap-2">
-                  <span>Google Gemini API Key</span>
-                  {brainSettings?.api_keys.has_gemini && (
-                    <CheckCircle2 className="h-4 w-4 text-green-500" />
-                  )}
-                </Label>
-                <div className="flex gap-2">
-                  <div className="relative flex-1">
-                    <Input
-                      id="gemini-key"
-                      type={visibleKeys.gemini ? 'text' : 'password'}
-                      placeholder={
-                        brainSettings?.api_keys.has_gemini
-                          ? brainSettings.api_keys.gemini_key
-                          : 'AIza-...'
-                      }
-                      value={localKeys.gemini || ''}
-                      onChange={(e) =>
-                        setLocalKeys({ ...localKeys, gemini: e.target.value || '' })
-                      }
-                      className="pr-20"
-                    />
-                    <Button
-                      type="button"
-                      variant="ghost"
-                      size="sm"
-                      className="absolute right-1 top-1/2 h-7 -translate-y-1/2 px-2"
-                      onClick={() =>
-                        setVisibleKeys({ ...visibleKeys, gemini: !visibleKeys.gemini })
-                      }
-                    >
-                      {visibleKeys.gemini ? (
-                        <EyeOff className="h-4 w-4" />
-                      ) : (
-                        <Eye className="h-4 w-4" />
-                      )}
-                    </Button>
-                  </div>
-                  <Button
-                    size="sm"
-                    onClick={() => handleSaveAPIKey('gemini', localKeys.gemini || '')}
-                    disabled={!localKeys.gemini || isSaving}
-                  >
-                    Save
-                  </Button>
-                </div>
-              </div>
-
-              {/* Ollama Base URL */}
-              <div className="space-y-3">
-                <Label htmlFor="ollama-url">Ollama Base URL</Label>
-                <div className="flex gap-2">
-                  <Input
-                    id="ollama-url"
-                    type="text"
-                    placeholder="http://localhost:11434"
-                    value={localOllamaUrl}
-                    onChange={(e) => setLocalOllamaUrl(e.target.value)}
-                  />
-                  <Button
-                    size="sm"
-                    onClick={() => handleSaveOllamaUrl(localOllamaUrl)}
-                    disabled={isSaving}
-                  >
-                    Save
-                  </Button>
-                </div>
-                <p className="text-xs text-muted-foreground">
-                  For local models using Ollama. Leave default unless running Ollama on
-                  a different host.
-                </p>
-              </div>
-            </CardContent>
-          </Card>
+          <ApiKeyPanel
+            brainSettings={brainSettings}
+            brainSettingsLoading={brainSettingsLoading}
+            isSaving={isSaving}
+            onSaveComplete={handleRefetchAll}
+            onSavingChange={setIsSaving}
+          />
         </TabsContent>
 
-        {/* BRAIN-033: RAG Settings Tab */}
+        {/* RAG Settings Tab */}
         <TabsContent value="rag-settings" className="space-y-6">
-          <Card>
-            <CardHeader>
-              <CardTitle className="flex items-center gap-2">
-                <Database className="h-5 w-5" />
-                RAG Configuration
-              </CardTitle>
-              <CardDescription>
-                Configure retrieval-augmented generation settings for knowledge base
-                queries.
-              </CardDescription>
-            </CardHeader>
-            <CardContent className="space-y-6">
-              {/* Enable RAG */}
-              <div className="flex items-center justify-between rounded-lg border p-4">
-                <div>
-                  <Label htmlFor="rag-enabled" className="text-base">
-                    Enable RAG
-                  </Label>
-                  <p className="text-sm text-muted-foreground">
-                    Use knowledge base context when generating responses
-                  </p>
-                </div>
-                <Switch
-                  id="rag-enabled"
-                  checked={brainSettings?.rag_config.enabled ?? true}
-                  onCheckedChange={(checked) =>
-                    handleSaveRAGConfig({ enabled: checked })
-                  }
-                />
-              </div>
-
-              {/* Chunk Settings */}
-              <div className="space-y-4">
-                <h3 className="font-semibold">Chunking Settings</h3>
-                <div className="grid grid-cols-2 gap-4">
-                  <div className="space-y-2">
-                    <Label htmlFor="chunk-size">Chunk Size</Label>
-                    <Input
-                      id="chunk-size"
-                      type="number"
-                      min={100}
-                      max={10000}
-                      value={brainSettings?.rag_config.chunk_size ?? 500}
-                      onChange={(e) =>
-                        handleSaveRAGConfig({
-                          chunk_size: parseInt(e.target.value, 10),
-                        })
-                      }
-                    />
-                    <p className="text-xs text-muted-foreground">Default: 500 tokens</p>
-                  </div>
-                  <div className="space-y-2">
-                    <Label htmlFor="chunk-overlap">Chunk Overlap</Label>
-                    <Input
-                      id="chunk-overlap"
-                      type="number"
-                      min={0}
-                      max={1000}
-                      value={brainSettings?.rag_config.chunk_overlap ?? 50}
-                      onChange={(e) =>
-                        handleSaveRAGConfig({
-                          chunk_overlap: parseInt(e.target.value, 10),
-                        })
-                      }
-                    />
-                    <p className="text-xs text-muted-foreground">Default: 50 tokens</p>
-                  </div>
-                </div>
-              </div>
-
-              {/* Retrieval Settings */}
-              <div className="space-y-4">
-                <h3 className="font-semibold">Retrieval Settings</h3>
-                <div className="grid grid-cols-2 gap-4">
-                  <div className="space-y-2">
-                    <Label htmlFor="max-chunks">Max Chunks</Label>
-                    <Input
-                      id="max-chunks"
-                      type="number"
-                      min={1}
-                      max={50}
-                      value={brainSettings?.rag_config.max_chunks ?? 5}
-                      onChange={(e) =>
-                        handleSaveRAGConfig({
-                          max_chunks: parseInt(e.target.value, 10),
-                        })
-                      }
-                    />
-                    <p className="text-xs text-muted-foreground">
-                      Maximum chunks to retrieve
-                    </p>
-                  </div>
-                  <div className="space-y-2">
-                    <Label htmlFor="score-threshold">Score Threshold</Label>
-                    <Input
-                      id="score-threshold"
-                      type="number"
-                      min={0}
-                      max={1}
-                      step={0.1}
-                      value={brainSettings?.rag_config.score_threshold ?? 0}
-                      onChange={(e) =>
-                        handleSaveRAGConfig({
-                          score_threshold: parseFloat(e.target.value),
-                        })
-                      }
-                    />
-                    <p className="text-xs text-muted-foreground">
-                      Minimum relevance score (0-1)
-                    </p>
-                  </div>
-                </div>
-              </div>
-
-              {/* Hybrid Search Weight */}
-              <div className="space-y-2">
-                <Label htmlFor="hybrid-weight">Hybrid Search Weight</Label>
-                <div className="flex items-center gap-4">
-                  <Input
-                    id="hybrid-weight"
-                    type="number"
-                    min={0}
-                    max={1}
-                    step={0.1}
-                    className="flex-1"
-                    value={brainSettings?.rag_config.hybrid_search_weight ?? 0.7}
-                    onChange={(e) =>
-                      handleSaveRAGConfig({
-                        hybrid_search_weight: parseFloat(e.target.value),
-                      })
-                    }
-                  />
-                  <span className="whitespace-nowrap text-sm text-muted-foreground">
-                    Vector:{' '}
-                    {Math.round(
-                      (brainSettings?.rag_config.hybrid_search_weight ?? 0.7) * 100
-                    )}
-                    % / BM25:{' '}
-                    {Math.round(
-                      (1 - (brainSettings?.rag_config.hybrid_search_weight ?? 0.7)) *
-                        100
-                    )}
-                    %
-                  </span>
-                </div>
-                <p className="text-xs text-muted-foreground">
-                  Balance between semantic (vector) and keyword (BM25) search
-                </p>
-              </div>
-
-              {/* Context Token Limit */}
-              <div className="space-y-2">
-                <Label htmlFor="token-limit">Context Token Limit</Label>
-                <Input
-                  id="token-limit"
-                  type="number"
-                  min={100}
-                  max={100000}
-                  value={brainSettings?.rag_config.context_token_limit ?? 4000}
-                  onChange={(e) =>
-                    handleSaveRAGConfig({
-                      context_token_limit: parseInt(e.target.value, 10),
-                    })
-                  }
-                />
-                <p className="text-xs text-muted-foreground">
-                  Maximum tokens for retrieved context (default: 4000)
-                </p>
-              </div>
-
-              {/* Include Sources */}
-              <div className="flex items-center justify-between rounded-lg border p-4">
-                <div>
-                  <Label htmlFor="include-sources" className="text-base">
-                    Include Source Citations
-                  </Label>
-                  <p className="text-sm text-muted-foreground">
-                    Add source references to retrieved context
-                  </p>
-                </div>
-                <Switch
-                  id="include-sources"
-                  checked={brainSettings?.rag_config.include_sources ?? true}
-                  onCheckedChange={(checked) =>
-                    handleSaveRAGConfig({ include_sources: checked })
-                  }
-                />
-              </div>
-            </CardContent>
-          </Card>
+          <RagConfigPanel
+            brainSettings={brainSettings}
+            isSaving={isSaving}
+            onSaveComplete={handleRefetchAll}
+            onSavingChange={setIsSaving}
+          />
         </TabsContent>
       </Tabs>
     </div>
