+++ frontend/src/features/routing/components/panels/index.ts
@@ -1,0 +1,12 @@
+/**
+ * Brain Settings Panel Components
+ *
+ * Modular panel components for the Brain Settings page.
+ * Each panel handles a specific area of configuration.
+ */
+
+export { ApiKeyPanel } from './ApiKeyPanel';
+export { ModelConfigPanel } from './ModelConfigPanel';
+export { RagConfigPanel } from './RagConfigPanel';
+export { RealtimeUsagePanel } from './RealtimeUsagePanel';
+export { TokenAnalyticsPanel } from './TokenAnalyticsPanel';
