+++ scripts/ralph/progress.txt
@@ -52,6 +52,28 @@
 
 ## Progress Log
 
+## 2025-02-26 11:30 - PREP-004
+- **What was implemented**: Decomposed BrainSettingsPage from 2109 lines into 5 modular panel components
+- **Files changed**:
+  - Created: frontend/src/features/routing/components/panels/index.ts
+  - Created: frontend/src/features/routing/components/panels/RealtimeUsagePanel.tsx (270 lines)
+  - Created: frontend/src/features/routing/components/panels/ApiKeyPanel.tsx (352 lines)
+  - Created: frontend/src/features/routing/components/panels/RagConfigPanel.tsx (261 lines)
+  - Created: frontend/src/features/routing/components/panels/ModelConfigPanel.tsx (643 lines)
+  - Created: frontend/src/features/routing/components/panels/TokenAnalyticsPanel.tsx (587 lines)
+  - Updated: frontend/src/features/routing/components/BrainSettingsPage.tsx (275 lines, down from 2109)
+- **Learnings for future iterations:**
+  - React components should be <300 lines; complex panels can be ~500-650 lines
+  - Panel components receive shared state via props (isSaving, onSaveComplete, onSavingChange)
+  - Use TypeScript interfaces for panel props to ensure type safety
+  - The main page component now orchestrates panels and handles data fetching
+  - Tabs reduced from 7 to 4 by combining related functionality in ModelConfigPanel
+  - Export type definitions from main component for use in panels
+
+---
+
+## Progress Log
+
 ## 2025-02-26 10:30 - PREP-002
 - **What was implemented**: Completed schema modularization - verified world_schemas.py and knowledge_schemas.py exist
 - **Files changed**:
